# docs/api/commerceflow-openapi.yaml
...
openapi: 3.0.3
info:
  title: CommerceFlow API
  version: 1.0.0
  description: |
    Microservices-based e-commerce order processing.

    Placing an order is **asynchronous**. `POST /api/orders` returns `201` with status `CREATED`
    as soon as the order is accepted and the orchestrator has been started; stock
    reservation, payment and notification then run over Kafka. Poll `GET /api/orders/{id}` to
    follow it — the status walks
    `CREATED → INVENTORY_RESERVED → PAID → COMPLETED`, or moves to `CANCELLED` with
    `failedStep` and `failureReason` telling you where and why it stopped.

    Every response is wrapped in a uniform envelope: `ApiResponse` on success, `ErrorResponse`
    on failure. Both carry the `correlationId`, which is the id to quote in a support request —
    it stitches the whole saga together in the logs.
  contact:
    name: CommerceFlow Platform Team
    email: platform@commerceflow.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local API Gateway (docker compose)
  - url: https://api.commerceflow.io
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/coupons/preview:
    post:
      tags: [Coupons]
      summary: What would this code be worth?
      description: |
        Validates a code against a basket **without spending a redemption**. A basket page that
        claimed one every time somebody typed a code would exhaust a limited campaign on curiosity.

        The consequence is stated rather than hidden: a code can pass here and be refused at
        checkout, because the last one went in between. That is what a limited campaign means.

        The saving comes back as **two figures**. A reduction on the goods lowers the tax with it;
        free delivery does not touch the tax at all. Collapsing them into one number is how a
        basket page ends up disagreeing with the checkout for reasons no customer can work out.

        A code that cannot be used is refused with `422` and a message naming the reason — expired,
        not started, fully redeemed, basket too small, already used. "Invalid code" for all of them
        leaves somebody unable to tell a typo from an expiry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, basketAmount]
              properties:
                code: { type: string, maxLength: 40, example: WELCOME10 }
                basketAmount: { type: number, description: The goods at list price }
                shippingAmount: { type: number }
                currency: { type: string, example: EUR }
      responses:
        "200":
          description: What it is worth
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/CouponPreview" }
        "422": { $ref: "#/components/responses/Conflict" }
  /api/coupons:
    get:
      tags: [Coupons]
      summary: List codes
      description: Administrator only. Newest first.
      parameters:
        - { name: active, in: query, schema: { type: boolean } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200": { description: A page of codes }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Coupons]
      summary: Create a code
      description: |
        Administrator only.

        `value` is a **fraction** for a percentage: `0.10` is ten per cent. A fixed amount is an
        amount in `currency`, and needs one — without it there is no way to tell whether "10 off"
        means euros or dong.

        A percentage of 1 or more is **refused rather than interpreted**. Accepting `10` and
        guessing which was meant is the kind of convenience that eventually creates a code worth
        ten times what somebody intended.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CouponRequest" }
      responses:
        "201": { description: The new code }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/coupons/{id}:
    put:
      tags: [Coupons]
      summary: Edit a code
      description: |
        Administrator only. Everything but the code itself, which is immutable: it is printed on
        posters and recorded as text on every order that used it, so renaming it makes all of that
        refer to something that no longer exists.

        The value, the window and the caps **are** editable on a live campaign — one doing better
        or worse than expected gets adjusted, and refusing that sends operators to the database.
        Changes affect the next order and nothing already placed.

        Lowering `maxRedemptions` below what has already been redeemed is refused with the actual
        number, rather than arriving as a constraint violation.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CouponRequest" }
      responses:
        "200": { description: The updated code }
        "422": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Coupons]
      summary: Delete an unused code
      description: >-
        Administrator only. Refused with `409` once it has been redeemed — those rows are the record
        of what the campaign cost. Switch it off instead, which stops it working immediately and
        keeps the history.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/coupons/{id}/redemptions:
    get:
      tags: [Coupons]
      summary: Who used a code
      description: >-
        Administrator only. `discountAmount` is what it was worth on that order, which can be less
        than the code's face value when the basket was smaller than the discount. That figure, not
        the value times the count, is what the campaign cost.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Every use, most recent first }

  # ------------------------------------------------------------------ delivery
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned by `POST /api/auth/login`

  # This file is the hand-written contract for the paths that matter most. It is deliberately
  # not exhaustive: each service serves its own generated definition at /v3/api-docs, aggregated
  # behind the gateway at /swagger-ui.html, and that is authoritative for everything else --
  # notably the back-office endpoints for warehouses, stock levels, product images and review
  # moderation, which no external client calls.
  responses:
...
    Forbidden:
      description: Authenticated, but not allowed (`FORBIDDEN`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ValidationFailed:
      description: Request validation failed; `fieldErrors` lists every violated constraint
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: |
        The request is well formed but the resource is not in a state that allows it
        (`ORDER_NOT_MODIFIABLE`, `CONFLICT`). The message says what to do about it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    CouponRequest:
      type: object
      required: [code, type, value]
      properties:
        code:
          type: string
          maxLength: 40
          description: Immutable once created. Stored and matched upper case.
        description: { type: string, maxLength: 200 }
        type: { type: string, enum: [PERCENTAGE, FIXED_AMOUNT, FREE_SHIPPING] }
        value:
          type: number
          description: >-
            A fraction for PERCENTAGE (0.1000 is 10%), an amount for FIXED_AMOUNT, ignored for
            FREE_SHIPPING. A percentage of 1 or more is refused rather than interpreted.
        currency:
          type: string
          description: >-
            Required for FIXED_AMOUNT. Such a code is refused on an order in another currency
            rather than converted — inventing an exchange rate at checkout is worse than refusing
            the code.
        minimumBasket: { type: number }
        maxRedemptions:
          type: integer
          nullable: true
          description: Null for unlimited. Cannot be lowered below what has already been redeemed.
        perCustomerLimit: { type: integer, nullable: true }
        validFrom: { type: string, format: date-time }
        validUntil: { type: string, format: date-time }
        active: { type: boolean }
    CouponPreview:
      type: object
      description: >-
        Two figures, not one: a reduction on the goods lowers the tax with it, and free delivery
        does not touch the tax at all.
      properties:
        code: { type: string }
        type: { type: string, enum: [PERCENTAGE, FIXED_AMOUNT, FREE_SHIPPING] }
        description: { type: string, nullable: true }
        goodsDiscount: { type: number }
        shippingDiscount: { type: number }
        currency: { type: string }
...
    ApiResponse:
      type: object
      description: Uniform success envelope.
      properties:
        success: { type: boolean, example: true }
        data: { description: The payload }
        message: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        correlationId:
          type: string
          nullable: true
          description: Quote this in a support request; it stitches the saga together in the logs.
    ErrorResponse:
      type: object
      description: Uniform error envelope.
      properties:
        success: { type: boolean, example: false }
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 409 }
        error: { type: string, example: Conflict }
        code:
          type: string
          example: INSUFFICIENT_STOCK
          description: Machine readable error code; stable across releases.
        message: { type: string }
        path: { type: string, example: /api/orders }
        correlationId: { type: string, nullable: true }
        fieldErrors:
          type: array
          nullable: true
          items:
            type: object
            properties:
              field: { type: string, example: email }
              rejectedValue: {}
              message: { type: string, example: must be a valid address }
...